// src/client/utils/Types.ts
export interface IPokemon {
    avg_spawns: number;
    candy: string;
    candy_count: number;
    egg: string;
    height: string;
    id: number;
    img: string;
    multipliers: Array<number>;
    name: string;
    next_evolution: Array<any>;
    num: string;
    spawn_chance: number;
    spawn_time: string;
    type: Array<string>;
    weaknesses: Array<string>;
    weight: string;
  }